[RSPEC COVERAGE FOR NON JS | DYNAMIC LISTS] clean up
